requirements.txt
pandas
GEOparse
biopython
streamlit
bs4